backend/requirements.txt
fastapi==0.104.1
python-socketio==5.11.1
uvicorn==0.24.0.post1
websockets==12.0
python-socketio[asgi]
